only create 1 Flask app
